+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to deser are documented here.
 
+## 0.8.0
+
+- Removed `for_each_event`.
+
 ## 0.7.0
 
 - Added support for `Box<T>`.